Add python additional paths for research
