+++ Queues/JobQueue.cs
@@ -98,7 +98,7 @@ namespace QuantConnect.Queues
         /// <returns></returns>
         public AlgorithmNodePacket NextJob(out string location)
         {
-            ConfigurePythonPaths();
+            PythonInitializer.ConfigurePythonPaths();
 
             location = GetAlgorithmLocation();
 
@@ -255,24 +255,6 @@ namespace QuantConnect.Queues
                 Console.WriteLine("Engine.Main(): Press any key to continue.");
                 System.Console.Read();
             }
-        }
-
-        /// <summary>
-        /// Gets the python additional paths from the config and adds them to Python using the PythonInitializer
-        /// </summary>
-        private static void ConfigurePythonPaths()
-        {
-            var pythonAdditionalPaths = Config.GetValue("python-additional-paths", Enumerable.Empty<string>());
-            PythonInitializer.AddPythonPaths(pythonAdditionalPaths.Where(path =>
-            {
-                var pathExists = Directory.Exists(path);
-                if (!pathExists)
-                {
-                    Log.Error($"JobQueue.ConfigurePythonPaths(): Unable to find python path: {path}. Skipping.");
-                }
-
-                return pathExists;
-            }));
         }
     }
 }
